Ignore when we use **kwargs, but not named kwds
When this is removed, an "incompatible type" error message shows up.
I decided to just follow how spaCy does this in the following lines of
code:
- https://github.com/explosion/spaCy/blob/c90dd6f265e5cef84de3a2a935728898e20b94df/spacy/scorer.py#L134
- https://github.com/explosion/spaCy/blob/4c058eb40a1843191352a1501bead0dc99526bed/spacy/language.py#L1164

I'm cutting corners here so I want to document it properly. For now, if
we want to fully-solve this error, there are two ways:

1. I can just pass each parameter to the row() function instead of using
the **settings dictionary. I decided not to do this because I find the
**settings approach to be elegant (I don't want to sacrifice neatness
for static type correctness).

2. We can use TypedDicts. However, I don't see this approach being done
in spaCy, and I might need to refactor the code for a bit so this might
take more time.

Let me know if there's something obvious that I missed!