Migrate codebase to support py36 and above (#22)
* Remove CI checks for py27

* Add type hints and remove py27 compat

In this commit, I added type hints to all public-facing
methods of the API. In addition, I also removed any traces of python2.7
compatibility (e.g., print_function, unicode checks, etc.)

For most of the docstrings, I updated 'unicode' -> 'str' just so that
it's clearer.

* Fix typing errors for diff_strings

* Ignore when we use **kwargs, but not named kwds

When this is removed, an "incompatible type" error message shows up.
I decided to just follow how spaCy does this in the following lines of
code:
- https://github.com/explosion/spaCy/blob/c90dd6f265e5cef84de3a2a935728898e20b94df/spacy/scorer.py#L134
- https://github.com/explosion/spaCy/blob/4c058eb40a1843191352a1501bead0dc99526bed/spacy/language.py#L1164

I'm cutting corners here so I want to document it properly. For now, if
we want to fully-solve this error, there are two ways:

1. I can just pass each parameter to the row() function instead of using
the **settings dictionary. I decided not to do this because I find the
**settings approach to be elegant (I don't want to sacrifice neatness
for static type correctness).

2. We can use TypedDicts. However, I don't see this approach being done
in spaCy, and I might need to refactor the code for a bit so this might
take more time.

Let me know if there's something obvious that I missed!

* Fix tables typing

* Fix dictionary getters by casting

Right now, functions such as icon, color, bg_color
have different types (str, int, etc.), and mypy gets confused
as to which one to use when we're doing some_dict.get(color, color)

What I did is to cast everything into string. One option is to do
overloads, but I don't think it's necessary for this method.

* Add typing_extensions to use Literal

* Run isort on all files

* Install dependencies in a separate step

* Remove unnecessary imports

* Add mypy to CI

* Separate steps for pytest and mypy

* Remove debugging lines

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* Fix mypy error when using append

As it turns out, the append method of a list expects a Tuple[Any],
it cannot be a List. I typecasted the argument's input to the
appropriate type.

* Update markdown.py typing to support iterable

Support Iterable when necessary. Be more explicit when using List
vs. Sequence. Right now, I see this as:

- Sequence: ordered, Collections: unordered

* Improve docstrings and typing

* Update typing for markdown.py

* Use general types like Collection inplace of Lists

* Improve docstrings

* Improve docstrings for optional values

* Fix typing on tables.py

* Initial fix on tables.py typing

* Fix optional param for traceback printer

* Update util.py typing and docstring

* Run isort on the codebase

* Remove future compat code in test suite

* Update wasabi/__init__.py

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* Remove encoding altogether

* Add quick updates on typing

* Update docstrings to reflect the actual type

Instead of depending on the docstring's text when specifying the type,
we specify the actual type as seen in the parameter.

* Make importing zip_longest make sense

* Update setup to include python_requires

* Remove unused imports

* Fix implicit optional to tables.py

* Fix implicit imports in printer.py

* Update wasabi/tables.py

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* Make typing for aligns param more explicit

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>