Add support for ANSI on Windows
This is a belated replacement for the reverted #24.

Improvements this time around:

Moved the relevant code into upstream colorama, so it plays nicely with
other packages using terminal features on Windows, like tqdm.

This also gives us basic ANSI support on any old Windows installs out
there that still don't have the native ANSI support. I think they're all
EOL already so not a big deal, but it's a free bonus, so why not.

Last time, we had to revert due to a bug in handling "fake" consoles
like Jupyter notebooks. This time, we're using code that's already been
out in the wild for ~1 month with no issues reported, so hopefully there
won't be any need for emergency reverts.